requirements.txt
python-hcl2
numpy
scikit-learn
gitpython
joblib